requirements.txt
aiogram==3.*
instaloader==4.*
python-dotenv==1.*
